platform: t186: fix argument list in aowake
bug 1686531

Reviewed-on: http://git-master/r/806198
(cherry picked from commit c601338162cd72d049f5aac76e18de60afb1d706)
Reviewed-on: http://git-master/r/815596
Tested-by: Timo Alho <talho@nvidia.com>
GVS: Gerrit_Virtual_Submit

+++ drivers/platform/tegra/tegra186-aowake.h
@@ -15,7 +15,7 @@
 #define TEGRA186_AOWAKE_H
 
 extern unsigned long tegra_aowake_read(unsigned int reg_offset);
-extern int tegra_aowake_write(unsigned int val, unsigned long reg_offset);
+extern int tegra_aowake_write(unsigned long val, unsigned int reg_offset);
 extern int tegra_aowake_update(unsigned int reg_offset, unsigned long mask,
 		unsigned long val);
 #ifdef CONFIG_PM_SLEEP
